Refactoring of fieldDef and column report object.
ReportDataStore has columns array with one column object for each column in a grid. Aggrid expects the cols to have
a certain properties some of these are part of the field information and some part of the report definition.

For ease of code reuse and aggrid agnostic the field definition associated with the column from the report's table field
structure is retained as a property on the column object as a whole in the `fieldDef` property.
This property mirrors the server field structure and can be used with the common validation utility on the node server
side as well as client side validation for report inline editing and possibly form field validation.

Beside the fieldDef object we add the `field` member to column object which is the name of the property to use
to access that columns data from the data rows hash for aggrid. Aggrid also needs `headerName` to title the column
header in the grid header row. The `order` property is the index sequence of the column

Some other values from fieldDef are elevated for convenience to column definition
level also `id, fieldType, fieldName, defaultValue, choices` i.e.

           field: fieldDef.name
           id: fieldDef.id
           fieldType: fieldDef.type
           fieldName: fieldDef.name
           defaultValue: fieldDef.defaultValue
           choices: fieldDef.multipleChoice.choices
           fieldDef: {} see below

The fields Definition is included as the `fieldDef` member. fieldDef has:

          fieldDef:  {
            // all fields have this information
            "id": 6,
            "type": "SCALAR", // generated fields (summary formula) are not scalar
            "name": "Brand",
            "builtIn": false,
            "dataIsCopyable": true,
            "includeInQuickSearch": true,
            "appearsByDefault": true,
            "userEditableValue": true,
            "required": false,
            "unique": false,
            "indexed": false,
            // only scalar fields have an optional defaultValue
            "defaultValue": {},
            "datatypeAttributes": {
              "type": "TEXT",
              // contents of clientSideAttributes varies by dataType
              "clientSideAttributes": {
                "bold": false,
                "width": 50,
                "word_wrap": true,
                "num_lines": 1,
                "max_chars": 0
              },
              "htmlAllowed": false
            },
            // only when field is multiple choice
            "multipleChoice": {
              "choices": []
              }
          }

--