CP-665 added a drawer to house the fields to add
When a user selects add column, a drawer will appear on the right with
a list of fields that they can add to the table.
When a field name is clicked, the column is added to the table in the
correct location based on which column dropdown they selected from.
The drawer will then close.

— added min-height: 300px because when the grid is displayed inside the
drawer, the dropdown gets cut off at the bottom of the records list