RED-951 Show spinner when retrieving users (#876)
Both spinners on the page (left nav and grid) checked the
RequestContext's isFetching status to confirm whether or not to show the
spinner.  This causes issues if the RequestContext request has completed
but the request to retrieve account users has not.

With this change the spinner on the account grid will only go away whent
he request to retrieve users has completed.