RED-968 Ticket Federation Redirect Fix for Firefox
Firefox has differing behavior with how it resolved promises when a
redirect is executing. When a redirect is executing the promise
generated from constructRedirectURL will reject() and return an error
object with a bunch of null fields.  If we encounter the object with
null fields, we will just swallow the error and continue along.