Wrap fields api assertions in promises
If an assert fails during an asynchronous test and that assert falls outside of a
promise, then the test will fail due to a timeout because the original promise never resolves.
Wrapping all asserts in promises ensures that when a test fails it will output a friendly
error message instead of timing out.